requirements.txt
twitchio
python-socketio
